requirements.txt
pandas>=2.0.0
requests>=2.31.0
rapidfuzz>=3.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
